Data types adjusted to int

+++ include/Mqtt/MeasurementsMqttMessagePublisher.h
@@ -12,7 +12,7 @@ class MeasurementsMqttMessagePublisher
 public:
     const String PubTopic = "remotecardiagz/pids/";
     MeasurementsMqttMessagePublisher(activatedPidsKeyValuePair *activePids);
-    void publishMessage(String topic, uint16_t value);
+    void publishMessage(String topic, int value);
 
 private:
     Ticker _mqttReconnectTimer;
